added firebase login, signup, example todo component
Switched from amplify to firebase. Installed firebase npm
package that enables interaction with firebase features. Set
up firebase login and signup methods, and wrapped them in
simple SignIn and SignUp components. Built an example todo app for
experimenting with firebase's key-value database structure and its
ability to keep track of updates to the database realtime and show an
up-to-date version of the data on the frontend.