added Post and PostCreator components
Added the Post component, which takes in a post's id and the auth
user's id, and renders that post. The Post component shows the
username of the user who created the post, the image associated with
the post, how many likes the post has, and whether or not the auth
user liked the post. Added the PostCreator component, which takes in
the auth user's id, and allows the user to create a new post. The
process of creating a post only requires uploading an image as of now.
With both of these components, there is not yet styling or any attempt
at making things pretty.