changed posts styling, added search page, modified profile pages
Changed posts and feed styling so that they rely more on the icons.
Added MainFeed component that allows customizing the header for the
main feed page. Separated profile component into CurrentUserProfile
and OtherUserProfile to allow different headers and props for the two
pages. The CurrentUserProfile operates the same as before, with an
additional header to distinguish it from the main feed page. The
OtherUserProfile component uses the url parameter uid to display the
posts for the user specified by the uid. Added search page which
allows entering a username to search for a user. It searches by
requesting all users in the database, and using a local string
matching function to only show entries that are similar to the search
text entered.